% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ca.census.tract.regions}
\alias{ca.census.tract.regions}
\title{Metadata about US Census Tracts map}
\usage{
data(ca.census.tract.regions)
}
\description{
The metadata is the tract id and the numeric county fips code.
}
\seealso{
?ca.census.tract.regions
}

