% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ca.tract.map}
\alias{ca.tract.map}
\title{Map of US Census Tracts in California}
\usage{
data(ca.tract.map)
}
\description{
The map comes from the 2010 Cartographic Boundary Shapefiles - Census Tracts 
from the US Census Bureau website.
A column named "region" was added which is a duplicate of the column GEOID. 
The GEOID is the concatenation of the 5-character county FIPS code and the 
6-character TRACTCE column.
A column named "county.fips.numeric" was added which is the numeric version
of the county FIPS code of the tract.
}
\references{
Taken from \url{https://www.census.gov/geo/maps-data/data/cbf/cbf_tracts.html}.
}
\seealso{
\link{ca.tract.regions}
}

