% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ca.tract.regions}
\alias{ca.tract.regions}
\title{Metadata about the California Census Tracts map}
\usage{
data(ca.tract.regions)
}
\description{
In addition to the region/GEOID it contains the TRACTCE and county FIPS code
as both an integer and character (i.e. with and without the leading "0")
}
\seealso{
\link{ca.tract.map}
}

