% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ca_tract_choropleth.R
\name{ca_tract_choropleth}
\alias{ca_tract_choropleth}
\title{Create a choropleth of US Census Tracts in California}
\usage{
ca_tract_choropleth(df, title = "", legend = "", num_colors = 7,
  tract_zoom = NULL, county_zoom = NULL)
}
\arguments{
\item{df}{A data.frame with a column named "region" and a column named "value".  Elements in
the "region" column must exactly match how census tracts are labelled in in the "region" column in ?ca.tract.regions}

\item{title}{An optional title for the map.}

\item{legend}{An optional name for the legend.}

\item{num_colors}{The number of colors on the map. A value of 1
will use a continuous scale. A value in [2, 9] will use that many colors.}

\item{tract_zoom}{An optional vector of tracts to zoom in on. Elements of this vector must exactly
match the names of tracts as they appear in the "region" column of ?ca.tract.regions.}

\item{county_zoom}{An optional vector of county FIPS codes to zoom in on. Elements of this
vector must exactly match the names of counties as they appear in the "county.fips.numeric" column
of ?ca.tract.regions.}
}
\description{
Create a choropleth of US Census Tracts in California
}
\examples{
\dontrun{
data(df_pop_ca_tract)

ca_tract_choropleth(df_pop_ca_tract,
                    title  = "2012 California Census Tract\\n Population Estimates",
                    legend = "Population")

# zoom in on San Francisco, which has FIPS code 6075
ca_tract_choropleth(df_pop_ca_tract,
                    title  = "2012 San Francisco Census Tract\\n Population Estimates",
                    legend = "Population",
                    county_zoom = 6075)

 # 2013 per capita income estimate
 data(df_ca_tract_demographics)
 df_ca_tract_demographics$value = df_ca_tract_demographics$per_capita
 ca_tract_choropleth(df_ca_tract_demographics,
                     title       = "2013 San Francisco Census Tract\\n Per Capita Income",
                     legend      = "Income",
                     num_colors  = 1,
                     county_zoom = 6075)
}
}
\seealso{
\url{https://www.census.gov/geo/reference/gtc/gtc_ct.html} for more information on Census Tracts
}

