% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/acs.R
\name{ca_tract_choropleth_acs}
\alias{ca_tract_choropleth_acs}
\title{Create a choropleth map of California Census Tracts from ACS data}
\usage{
ca_tract_choropleth_acs(tableId, endyear = 2011, span = 5, num_colors = 7,
  tract_zoom = NULL, county_zoom = NULL)
}
\arguments{
\item{tableId}{The id of an ACS table}

\item{endyear}{The end year of the survey to use.  See acs.fetch (?acs.fetch) and http://1.usa.gov/1geFSSj for details.}

\item{span}{The span of time to use.  See acs.fetch and http://1.usa.gov/1geFSSj for details.}

\item{num_colors}{The number of colors on the map. A value of 1
will use a continuous scale. A value in [2, 9] will use that many colors.}

\item{tract_zoom}{An optional list of states to zoom in on. Must come from the "region" column in
?ca.tract.regions.}

\item{county_zoom}{An optional list of counties to zoom in on. Must match entries of the county.fips.numeric column of
?ca.tract.regions.}
}
\value{
A choropleth.
}
\description{
Creates a choropleth of California Census Tracts the US Census' American Community Survey (ACS) data.
Requires the acs package to be installed, and a Census API Key to be set with
the acs's api.key.install function.  Census API keys can be obtained at http://www.census.gov/developers/tos/key_request.html.
}
\references{
Uses the acs package created by Ezra Haber Glenn.
}
\seealso{
\code{api.key.install} in the acs package which sets an Census API key for the acs library

http://factfinder2.census.gov/faces/help/jsf/pages/metadata.xhtml?lang=en&type=survey&id=survey.en.ACS_ACS
which contains a list of all ACS surveys.
}
\keyword{acs}
\keyword{choropleth,}

