% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ca_tract_choropleth.R
\docType{data}
\name{CaTractChoropleth}
\alias{CaTractChoropleth}
\title{An R6 object for creating choropleths of California Census Tracts.}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x10fc51380> 
 - attr(*, "name")= chr "CaTractChoropleth_generator"
}}
\usage{
CaTractChoropleth
}
\description{
An R6 object for creating choropleths of California Census Tracts.
}
\keyword{datasets}

