% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_ca_tract_demographics}
\alias{df_ca_tract_demographics}
\title{A data.frame containing demographic statistics for each Census Tract in California.}
\usage{
data(df_ca_tract_demographics)
}
\description{
A data.frame containing demographic statistics for each Census Tract in California.
}
\references{
Data comes from the 2013 5-year American Community Survey (ACS). Data generated by
?get_ca_tract_demographics.
}
\keyword{data}

