% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_pop_ca_tract}
\alias{df_pop_ca_tract}
\title{A data.frame containing population estimates for California Census Tracts in 2012.}
\usage{
data(df_pop_ca_tract)
}
\description{
A data.frame containing population estimates for California Census Tracts in 2012.
}
\references{
Taken from the US American Community Survey (ACS) 5 year estimates.
}
\seealso{
?df_pop_ca_tract
}

