% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_ca_tract_demographics.R
\name{get_ca_tract_demographics}
\alias{get_ca_tract_demographics}
\title{Get a handful of demographic variables on California Census Tracts from the US Census Bureau as a data.frame.}
\usage{
get_ca_tract_demographics(endyear = 2013, span = 5)
}
\arguments{
\item{endyear}{The end year for the survey}

\item{span}{The span of the survey}
}
\description{
The data comes from the American Community Survey (ACS). The variables are: total population, percent White
not Hispanic, Percent Black or African American not Hispanic, percent Asian not Hispanic,
percent Hispanic all races, per-capita income, median rent and median age.
}
\examples{
\dontrun{
df = get_ca_tract_demographics(endyear=2010, span=5)
colnames(df)

# analyze the percent of people who are white not hispanic
# a boxplot shows the distribution
boxplot(df$percent_white)

# a choropleth map shows the location of the values in california
# set the 'value' column to be the column we want to render
df$value = df$percent_white
ca_tract_choropleth(df,
                    title="2010 Census Tracts\\nPercent White not Hispanic",
                    legend="Percent")

# zoom into san francisco county
ca_tract_choropleth(df,
                    title="2010 Census Tracts\\nPercent White not Hispanic",
                    legend="Percent",
                    county_zoom=6075)
}
}
\references{
The choroplethr guide to Census data: http://cran.r-project.org/web/packages/choroplethr/vignettes/e-mapping-us-census-data.html
}

