% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ca_tract_choropleth.R
\name{ca_tract_choropleth}
\alias{ca_tract_choropleth}
\title{Create a choropleth of US Census Tracts in California}
\usage{
ca_tract_choropleth(df, title = "", legend = "", num_colors = 7,
  tract_zoom = NULL, county_zoom = NULL)
}
\arguments{
\item{df}{A data.frame with a column named "region" and a column named "value".  Elements in
the "region" column must exactly match how census tracts are labelled in in the "region" column in ?ca.census.tract.regions}

\item{title}{An optional title for the map.}

\item{legend}{An optional name for the legend.}

\item{num_colors}{The number of colors on the map. A value of 1
will use a continuous scale. A value in [2, 9] will use that many colors.}

\item{tract_zoom}{An optional vector of tracts to zoom in on. Elements of this vector must exactly
match the names of tracts as they appear in the "region" column of ?ca.census.tract.regions.}

\item{county_zoom}{An optional vector of county FIPS codes to zoom in on. Elements of this
vector must exactly match the names of counties as they appear in the "county.fips.numeric" column
of ?ca.census.tract.regions.}
}
\description{
Create a choropleth of US Census Tracts in California
}
\seealso{
\url{https://www.census.gov/geo/reference/gtc/gtc_ct.html} for more information on Census Tracts
}

