% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs.R
\name{ca_tract_choropleth_acs}
\alias{ca_tract_choropleth_acs}
\title{Create a choropleth map of California Census Tracts from ACS data}
\usage{
ca_tract_choropleth_acs(tableId, endyear = 2011, span = 5, num_colors = 7,
  tract_zoom = NULL, county_zoom = NULL)
}
\arguments{
\item{tableId}{The id of an ACS table}

\item{endyear}{The end year of the survey to use.  See acs.fetch (?acs.fetch) and http://1.usa.gov/1geFSSj for details.}

\item{span}{The span of time to use.  See acs.fetch and http://1.usa.gov/1geFSSj for details.}

\item{num_colors}{The number of colors on the map. A value of 1 
will use a continuous scale. A value in [2, 9] will use that many colors.}

\item{tract_zoom}{An optional list of states to zoom in on. Must come from the "region" column in
?ca.tract.regions.}

\item{county_zoom}{An optional list of counties to zoom in on. Must match entries of the county.fips.numeric column of 
?ca.tract.regions.}
}
\value{
A choropleth.
}
\description{
Creates a choropleth of California Census Tracts the US Census' American Community Survey (ACS) data.  
Requires the acs package to be installed, and a Census API Key to be set with 
the acs's api.key.install function.  Census API keys can be obtained at http://www.census.gov/developers/tos/key_request.html.
}
\examples{
\dontrun{
# per capita income, san francisco (fips code 6075)
ca_tract_choropleth_acs("B19301", county_zoom=6075)

# aggregate public assistance income, san francisco
ca_tract_choropleth_acs("B19067", county_zoom=6075)
}
}
\references{
The choroplethr guide to Census data: http://cran.r-project.org/web/packages/choroplethr/vignettes/e-mapping-us-census-data.html

Uses the acs package created by Ezra Haber Glenn.
}
\seealso{
\code{api.key.install} in the acs package which sets an Census API key for the acs library
}
\keyword{acs}
\keyword{choropleth,}

