% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_tract_choropleth.R
\docType{data}
\name{CaTractChoropleth}
\alias{CaTractChoropleth}
\title{An R6 object for creating choropleths of California Census Tracts.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CaTractChoropleth
}
\description{
An R6 object for creating choropleths of California Census Tracts.
}
\keyword{datasets}

