% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_pop_ca_tract}
\alias{df_pop_ca_tract}
\title{A data.frame containing population estimates for California Census Tracts in 2012.}
\usage{
data(df_pop_ca_tract)
}
\description{
A data.frame containing population estimates for California Census Tracts in 2012.
}
\examples{
data(df_pop_ca_tract)

# zoom in on San Francisco, which has FIPS code 6075
ca_tract_choropleth(df_pop_ca_tract, 
                    title  = "2012 San Francisco Census Tract\\n Population Estimates", 
                    legend = "Population",
                    county_zoom = 6075)
\dontrun{
ca_tract_choropleth(df_pop_ca_tract, 
                    title  = "2012 California Census Tract\\n Population Estimates", 
                    legend = "Population")

}
}
\references{
Taken from the US American Community Survey (ACS) 5 year estimates.
}
\seealso{
\link{df_ca_tract_demographics}
}

